refacotor: dayicon 수정 및 날짜관련 format 유틸로 이동
