+++ frontend/.vscode/settings.json
@@ -4,5 +4,5 @@
         ["cva\\(([^)]*)\\)", "[\"'`]([^\"'`]*).*?[\"'`]"],
         ["cx\\(([^)]*)\\)", "(?:'|\"|`)([^']*)(?:'|\"|`)"]
     ],
-    "cSpell.words": ["fcst", "KAKAO", "khai", "Sido", "tanstack"]
+    "cSpell.words": ["fcst", "KAKAO", "khai", "normalleading", "Sido", "tanstack"]
 }
