+++ frontend/src/components/journals/CustomCalendar.tsx
@@ -6,14 +6,8 @@ import { queryStringKeys } from '@/constants';
 import PrevMonth from '@/assets/buttons/btn-prev-month.svg';
 import NextMonth from '@/assets/buttons/btn-next-month.svg';
 import { fetchDogMonthStatistic, period } from '@/api/dog';
-import { formDate, formDay } from '@/utils/date';
+import { formCalendar, formDate, formDay } from '@/utils/date';
 import useCalendar from '@/hooks/useCalendar';
-
-const formCalendar = (date: Date) => {
-    const year = date.getFullYear();
-    const month = date.getMonth() + 1;
-    return `${year}년 ${month}`;
-};
 
 export default function CustomCalendar() {
     const location = useLocation();
